+++ examples/tutorials/IK_motion_planning_grasp.py
@@ -1,4 +1,5 @@
 import genesis as gs
+from genesis.utils.path_planing import RRTConnect_OMPL
 import numpy as np
 
 ########################## init ##########################
@@ -59,8 +60,9 @@ qpos = franka.inverse_kinematics(
 )
 # gripper open pos
 qpos[-2:] = 0.04
-path = franka.plan_path(
-    qpos_goal=qpos,
+planner = RRTConnect_OMPL(franka)
+path = planner.plan(
+    qpos,
     num_waypoints=200,  # 2s duration
 )
 # draw the planned path
