+++ genesis/utils/path_planing.py
@@ -15,6 +15,8 @@ try:
     IS_OMPL_AVAILABLE = True
 
     __all__ = [
+        "RRTConnect",
+        "RRT",
         "RRTConnect_OMPL",
         "RRT_OMPL",
         "PRM_OMPL",
@@ -27,7 +29,10 @@ try:
 except ImportError:
     IS_OMPL_AVAILABLE = False
     
-    __all__ = []
+    __all__ = [
+        "RRTConnect",
+        "RRT",
+    ]
 
 
 
@@ -86,10 +91,10 @@ class PathPlanner(ABC):
         if check_joint_limits:
             # NOTE: process joint limit
             if (qpos_start < self.default_q_limit[0]).any() or (qpos_start > self.default_q_limit[1]).any():
-                gs.raise_exception_from("`qpos_start` exceeds joint limit.")
+                gs.logger.warning("`qpos_start` exceeds joint limit.")
 
             if (qpos_goal < self.default_q_limit[0]).any() or (qpos_goal > self.default_q_limit[1]).any():
-                gs.raise_exception_from("`qpos_goal` exceeds joint limit.")
+                gs.logger.warning("`qpos_goal` exceeds joint limit.")
         return qpos_goal, qpos_start
     
     def validate_input_qpos_batch(self, qpos_goal, qpos_start, envs_idx, *, check_joint_limits=True):
@@ -114,10 +119,10 @@ class PathPlanner(ABC):
         if check_joint_limits:
             # NOTE: process joint limit
             if (qpos_start < self.default_q_limit[0]).any() or (qpos_start > self.default_q_limit[1]).any():
-                gs.raise_exception_from("`qpos_start` exceeds joint limit.")
+                gs.logger.warning("`qpos_start` exceeds joint limit.")
 
             if (qpos_goal < self.default_q_limit[0]).any() or (qpos_goal > self.default_q_limit[1]).any():
-                gs.raise_exception_from("`qpos_goal` exceeds joint limit.")
+                gs.logger.warning("`qpos_goal` exceeds joint limit.")
         return qpos_goal, qpos_start
         
     @abstractmethod
@@ -127,7 +132,9 @@ class PathPlanner(ABC):
         qpos_start=None,
     ):...
 
-
+    # ------------------------------------------------------------------------------------
+    # -------------------------------path planing utils-----------------------------------
+    # ------------------------------------------------------------------------------------
     @ti.kernel
     def interpolate_path(
         self,
@@ -137,7 +144,7 @@ class PathPlanner(ABC):
         mask: ti.types.ndarray(),       # [B]
     ):
         ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
-        for i_b in range(path.shape[1]):
+        for i_b in range(path.shape[0]):
             if not mask[i_b]:
                 continue
             num_samples = sample_ind[i_b, 1] - sample_ind[i_b, 0]
@@ -147,18 +154,42 @@ class PathPlanner(ABC):
                 step = (end - start) / num_samples
                 for i_s in range(num_samples):
                     tensor[i_b, sample_ind[i_b, 0] + i_s, i_q] = start + step * i_s
-                    
+
     def check_collision(self, path, ignore_geom_pairs=None, envs_idx=None):
-        save_qpos = self._entity.get_qpos()
         path = path.transpose(1,0) # N, B, Dof
         res = torch.zeros(path.shape[1], dtype=gs.tc_int, device=gs.device)
         for qpos in path:
-            self._entity.set_qpos(qpos)
+            self._entity.set_qpos(qpos, envs_idx=envs_idx)
             self._solver._kernel_detect_collision()
             self._kernel_check_collision(res, ignore_geom_pairs, envs_idx)
-        self._entity.set_qpos(save_qpos)
         return res
-        
+    
+    @ti.func
+    def _func_check_collision(
+        self,
+        ignore_geom_pairs: ti.types.ndarray(),
+        i_b: ti.int32,
+    ):
+        collision_detected = False
+        for i_c in range(self._solver.collider.n_contacts[i_b]):
+            i_ga = self._solver.collider.contact_data[i_c, i_b].geom_a
+            i_gb = self._solver.collider.contact_data[i_c, i_b].geom_b
+
+            ignore = False
+            for i_p in range(ignore_geom_pairs.shape[0]):
+                if (ignore_geom_pairs[i_p, 0] == i_ga and ignore_geom_pairs[i_p, 1] == i_gb) or (
+                    ignore_geom_pairs[i_p, 0] == i_gb and ignore_geom_pairs[i_p, 1] == i_ga):
+                    ignore = True
+                    break
+            if ignore:
+                continue
+
+            # TODO: handle self-collision (except the case for closed gripper)
+            if (self.geom_start <= i_ga < self.geom_end) ^ (self.geom_start <= i_gb < self.geom_end):
+                collision_detected = True
+                break
+        return collision_detected
+    
     @ti.kernel
     def _kernel_check_collision(
         self,
@@ -168,27 +199,9 @@ class PathPlanner(ABC):
     ):
         ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
         for i_b in envs_idx:
-            collision_detected = False
-            for i_c in range(self._solver.collider.n_contacts[i_b]):
-                i_ga = self._solver.collider.contact_data[i_c, i_b].geom_a
-                i_gb = self._solver.collider.contact_data[i_c, i_b].geom_b
-
-                ignore = False
-                for i_p in range(ignore_geom_pairs.shape[0]):
-                    if (ignore_geom_pairs[i_p, 0] == i_ga and ignore_geom_pairs[i_p, 1] == i_gb) or (
-                        ignore_geom_pairs[i_p, 0] == i_gb and ignore_geom_pairs[i_p, 1] == i_ga):
-                        ignore = True
-                        break
-                if ignore:
-                    continue
-
-                # TODO: handle self-collision (except the case for closed gripper)
-                if (self.geom_start <= i_ga < self.geom_end) ^ (self.geom_start <= i_gb < self.geom_end):
-                    # collision detected
-                    collision_detected = True
-                    break
+            collision_detected = self._func_check_collision(ignore_geom_pairs, i_b)
             tensor[i_b] = tensor[i_b] or collision_detected
-
+            
     def shortcut_path(self, path_mask, path, iterations=50, ignore_geom_pairs=None, envs_idx=None):
         """
         path_mask: torch.Tensor
@@ -200,26 +213,16 @@ class PathPlanner(ABC):
         """
         path_mask = path_mask.t()  # B, N
         path = path.transpose(1,0) # B, N, Dof
-        for _ in range(iterations):
+        for i in range(iterations):
             ind = torch.multinomial(path_mask, 2).sort()[0] # B, 2
             ind_mask = (ind[:,1] - ind[:,0]) > 1
-            tmp_path = path.clone()
-            self.interpolate_path(tmp_path, path, ind, ind_mask)
+            tmp_path = path.clone().contiguous()
+            self.interpolate_path(tmp_path, path.contiguous(), ind.contiguous(), ind_mask)
             collision_mask = self.check_collision(tmp_path, ignore_geom_pairs, envs_idx) # B
             path[~collision_mask] = tmp_path[~collision_mask]
+            if i == 10:
+                break
         return path.transpose(1,0)
-
-    def smooth_path(self, path):
-        pass
-
-
-    def simplify_path(self, path):
-        """
-        Paramters
-        ---------
-        path: torch.Tensor
-            path in [N,B,Ndof]
-        """
 
 
 @ti.data_oriented
@@ -241,10 +244,616 @@ class RRT(PathPlanner):
         the maximum step size of qpos in randians (or in meters)
     """
 
+    def _init_rrt_fields(self, goal_bias=0.05, max_nodes=2000, pos_tol=5e-3, max_step_size=0.1):
+        self._rrt_goal_bias = goal_bias
+        self._rrt_max_nodes = max_nodes
+        self._rrt_pos_tol = pos_tol # .28 degree
+        self._rrt_max_step_size = max_step_size # NOTE: in radian (about 5.7 degree)
+        self._rrt_start_configuration = ti.field(dtype=gs.ti_float, shape=self._solver._batch_shape(self.n_qs))
+        self._rrt_goal_configuration = ti.field(dtype=gs.ti_float, shape=self._solver._batch_shape(self.n_qs))
+        self.struct_rrt_node_info = ti.types.struct(
+            configuration=ti.types.vector(self.n_qs, gs.ti_float),
+            parent_idx=gs.ti_int,
+        )
+        self._rrt_node_info = self.struct_rrt_node_info.field(shape=self._solver._batch_shape(self._rrt_max_nodes))
+        self._rrt_tree_size = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        self._rrt_is_active = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        self._rrt_goal_reached_node_idx = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        
+    def _reset_rrt_fields(self):
+        self._rrt_start_configuration.fill(0.0)
+        self._rrt_goal_configuration.fill(0.0)
+        self._rrt_node_info.parent_idx.fill(-1)
+        self._rrt_node_info.configuration.fill(0.0)
+        self._rrt_tree_size.fill(0)
+        self._rrt_is_active.fill(0)
+        self._rrt_goal_reached_node_idx.fill(-1)
+    
+    @ti.kernel
+    def _kernel_rrt_init(
+        self,
+        qpos_start: ti.types.ndarray(),
+        qpos_goal: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray()
+    ):
+        # NOTE: run IK before this
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b_ in range(envs_idx.shape[0]):
+            i_b = envs_idx[i_b_]
+            for i_q in ti.static(range(self.n_qs)):
+                # save original qpos
+                self._rrt_start_configuration[i_q, i_b] = qpos_start[i_b_, i_q]
+                self._rrt_goal_configuration[i_q, i_b] = qpos_goal[i_b_, i_q]
+                self._rrt_node_info[0, i_b].configuration[i_q] = qpos_start[i_b_, i_q]
+            self._rrt_node_info[0, i_b].parent_idx = 0
+            self._rrt_tree_size[i_b] = 1
+            self._rrt_is_active[i_b] = 1
+
+    @ti.kernel
+    def _kernel_rrt_cleanup(
+        self,
+        qpos_current: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray()
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b_ in range(envs_idx.shape[0]):
+            i_b = envs_idx[i_b_]
+            for i_q in ti.static(range(self.n_qs)):
+                self._solver.qpos[i_q + self._q_start, i_b] = qpos_current[i_b_, i_q]
+            self._solver._func_forward_kinematics_entity(self._entity._idx_in_solver, i_b)
+    
+    @ti.kernel
+    def _kernel_rrt_step1(
+        self,
+        q_limit_lower: ti.types.ndarray(),
+        q_limit_upper: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray(),
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b in envs_idx:
+            if not self._rrt_is_active[i_b]:
+                continue
+            
+            random_sample = ti.Vector([
+                q_limit_lower[i_q] + ti.random(dtype=gs.ti_float) * (q_limit_upper[i_q] - q_limit_lower[i_q])
+                for i_q in ti.static(range(self.n_qs))
+            ])
+            if ti.random() < self._rrt_goal_bias:
+                random_sample = ti.Vector([
+                    self._rrt_goal_configuration[i_q, i_b]
+                    for i_q in ti.static(range(self.n_qs))
+                ])
+
+            # find nearest neighbor
+            nearest_neighbor_idx = -1
+            nearest_neighbor_dist = gs.ti_float(1e30) 
+            ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+            for i_n in range(self._rrt_tree_size[i_b]):
+                dist = (self._rrt_node_info.configuration[i_n, i_b] - random_sample).norm()
+                if dist < nearest_neighbor_dist:
+                    nearest_neighbor_dist = dist
+                    nearest_neighbor_idx = i_n
+
+            # steer from nearest neighbor to random sample
+            nearest_config = self._rrt_node_info.configuration[nearest_neighbor_idx, i_b]
+            direction = random_sample - nearest_config
+            steer_result = ti.Vector.zero(gs.ti_float, self.n_qs)
+            for i_q in ti.static(range(self.n_qs)):
+                # If the step size exceeds max_step_size, clip it
+                if abs(direction[i_q]) > self._rrt_max_step_size:
+                    direction[i_q] = ti.math.sign(direction[i_q]) * self._rrt_max_step_size
+                steer_result[i_q] = nearest_config[i_q] + direction[i_q]
+
+            if self._rrt_tree_size[i_b] < self._rrt_max_nodes - 1:
+                # add new node
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].configuration = steer_result
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].parent_idx = nearest_neighbor_idx
+                self._rrt_tree_size[i_b] += 1
+
+                # set the steer result and collision check for i_b
+                for i_q in ti.static(range(self.n_qs)):
+                    self._solver.qpos[i_q + self._q_start, i_b] = steer_result[i_q]
+                self._solver._func_forward_kinematics_entity(self._entity._idx_in_solver, i_b)
+                self._solver._func_update_geoms(i_b)
+    
+    @ti.func
+    def _func_check_collision(
+        self,
+        ignore_geom_pairs: ti.types.ndarray(),
+        i_b: ti.int32,
+    ):
+        collision_detected = False
+        for i_c in range(self._solver.collider.n_contacts[i_b]):
+            i_ga = self._solver.collider.contact_data[i_c, i_b].geom_a
+            i_gb = self._solver.collider.contact_data[i_c, i_b].geom_b
+
+            ignore = False
+            for i_p in range(ignore_geom_pairs.shape[0]):
+                if (ignore_geom_pairs[i_p, 0] == i_ga and ignore_geom_pairs[i_p, 1] == i_gb) or (
+                    ignore_geom_pairs[i_p, 0] == i_gb and ignore_geom_pairs[i_p, 1] == i_ga):
+                    ignore = True
+                    break
+            if ignore:
+                continue
+
+            # TODO: handle self-collision (except the case for closed gripper)
+            if (self.geom_start <= i_ga < self.geom_end) ^ (self.geom_start <= i_gb < self.geom_end):
+                collision_detected = True
+                break
+        return collision_detected
+
+    @ti.kernel
+    def _kernel_rrt_step2(
+        self,
+        ignore_geom_pairs: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray(),
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b in envs_idx:
+            if not self._rrt_is_active[i_b]:
+                continue
+
+            collision_detected = self._func_check_collision(ignore_geom_pairs, i_b)
+            if collision_detected:
+                self._rrt_tree_size[i_b] -= 1
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].configuration = 0.0
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].parent_idx = -1
+            else:
+                # check the obtained steer result is within goal configuration only if no collision
+                flag = True
+                for i_q in range(self.n_qs):
+                    if abs(self._solver.qpos[i_q + self._q_start, i_b] - self._rrt_goal_configuration[i_q, i_b]) > self._rrt_pos_tol:
+                        flag = False
+                        break
+                if flag:
+                    self._rrt_goal_reached_node_idx[i_b] = self._rrt_tree_size[i_b] - 1
+                    self._rrt_is_active[i_b] = 0
+        
+    def plan(
+        self,
+        qpos_goal,
+        qpos_start=None,
+        resolution=0.01,
+        timeout=5.0,
+        max_retry=1,
+        smooth_path=True,
+        num_waypoints=100,
+        envs_idx=None,
+    ):
+        """
+        Plan a path from `qpos_start` to `qpos_goal`.
+        qpos_goal : array_like
+            The goal state. [B, Nq] or [1, Nq]
+        qpos_start : None | array_like, optional
+            The start state. If None, the current state of the rigid entity will be used. Defaults to None. [B, Nq] or [1, Nq]
+        """
+        import time
+        if self._solver.n_envs > 0:
+            envs_idx = self._solver._sanitize_envs_idx(envs_idx)
+        else:
+            envs_idx = torch.zeros(1, dtype=gs.tc_int, device=gs.device)
+
+        qpos_cur = self._entity.get_qpos()
+        if qpos_start is None:
+            qpos_start = qpos_cur.clone()
+        else:
+            if qpos_start.ndim == 1:
+                qpos_start = qpos_start.unsqueeze(0)
+            assert qpos_start.ndim == 2
+            if qpos_start.shape[0] == 1:
+                qpos_start = qpos_start.repeat(len(envs_idx), 1)
+
+        if qpos_goal.ndim == 1:
+            qpos_goal = qpos_goal.unsqueeze(0)
+        assert qpos_goal.ndim == 2
+        if qpos_goal.shape[0] == 1:
+            qpos_goal = qpos_goal.repeat(len(envs_idx), 1)
+        
+        self._init_rrt_fields()
+        self._reset_rrt_fields()
+        self._kernel_rrt_init(qpos_start.contiguous(), qpos_goal.contiguous(), envs_idx)
+
+        self._entity.set_qpos(qpos_start)
+        self._solver._kernel_detect_collision()
+        scene_contact_info = self._solver.collider.contact_data.to_torch(gs.device)
+        n_contacts = self._solver.collider.n_contacts.to_torch(gs.device)
+
+        valid_mask = torch.logical_or(
+            torch.logical_and(
+                scene_contact_info["geom_a"] >= self.geom_start,
+                scene_contact_info["geom_a"] < self.geom_end,
+            ),
+            torch.logical_and(
+                scene_contact_info["geom_b"] >= self.geom_start,
+                scene_contact_info["geom_b"] < self.geom_end,
+            ),
+        )
+        contact_indices = torch.arange(valid_mask.shape[0], device=valid_mask.device).unsqueeze(1)
+        valid_mask = torch.logical_and(valid_mask, contact_indices < n_contacts)
+
+        max_env_collisions = int(torch.max(n_contacts).item())
+        valid_mask = valid_mask[:max_env_collisions]
+        geom_a = scene_contact_info["geom_a"][:max_env_collisions][valid_mask] # N 
+        geom_b = scene_contact_info["geom_b"][:max_env_collisions][valid_mask] # N
+        assert len(geom_a) == len(geom_b)
+
+        # NOTE: we will reduce the contacts in batch dim assuming internal geom collisions are the same for a robot
+        stacked_tensors = torch.stack((geom_a, geom_b), dim=1)
+        unique_pairs = torch.unique(stacked_tensors, dim=0) # N', 2
+
+        gs.logger.info("start rrt planning...")
+        start = time.time()
+        for i_n in range(self._rrt_max_nodes):
+            if self._rrt_is_active.to_torch().any():
+                self._kernel_rrt_step1(
+                    q_limit_lower=self._entity.q_limit[0],
+                    q_limit_upper=self._entity.q_limit[1],
+                    envs_idx=envs_idx,
+                )
+                self._solver._kernel_detect_collision()
+                self._kernel_rrt_step2(
+                    ignore_geom_pairs=unique_pairs,
+                    envs_idx=envs_idx,
+                )
+            else:
+                break
+        
+        if self._rrt_is_active.to_torch().any():
+            gs.logger.warning(f"rrt planning failed in {self._rrt_is_active.to_torch().sum()} envs")
+        
+        gs.logger.info(f"rrt planning time: {time.time() - start}")
+        start = time.time()
+        ts = self._rrt_tree_size.to_torch(device=gs.device)[envs_idx]
+        g_n = self._rrt_goal_reached_node_idx.to_torch(device=gs.device)[envs_idx] # B
+
+        node_info = self._rrt_node_info.to_torch(device=gs.device)
+        parents_idx = node_info["parent_idx"][:, envs_idx]
+        configurations = node_info["configuration"][:, envs_idx]
+
+        res = [g_n]
+        for _ in range(ts.max()):
+            g_n = parents_idx[g_n, torch.arange(len(envs_idx))]
+            res.append(g_n)
+            if torch.all(g_n == 0):
+                break
+        res_idx = torch.stack(list(reversed(res)), dim=0)
+        sol = configurations[res_idx, torch.arange(len(envs_idx))] # N, B, DoF
+
+        sol = align_weypoints_length(sol, rrt_valid_mask(res_idx), len(res_idx))
+        sol = self.shortcut_path(
+            torch.ones_like(sol[...,0]), sol, iterations=10, ignore_geom_pairs=unique_pairs, envs_idx=envs_idx
+        )
+
+        # interpolate to make num_waypoints
+        sol = align_weypoints_length(sol, torch.ones_like(sol[...,0]).bool(), num_waypoints)
+        gs.logger.info(f"path post-processing time: {time.time() - start}")
+
+        self._kernel_rrt_cleanup(qpos_cur.contiguous(), envs_idx)
+        return sol
+
 
 @ti.data_oriented
 class RRTConnect(PathPlanner):
-    pass
+    def _init_rrt_connect_fields(self, goal_bias=0.05, max_nodes=2000, pos_tol=5e-3, max_step_size=0.1):
+        self._rrt_goal_bias = goal_bias
+        self._rrt_max_nodes = max_nodes
+        self._rrt_pos_tol = pos_tol # .28 degree
+        self._rrt_max_step_size = max_step_size # NOTE: in radian (about 5.7 degree)
+        self._rrt_start_configuration = ti.field(dtype=gs.ti_float, shape=self._solver._batch_shape(self.n_qs))
+        self._rrt_goal_configuration = ti.field(dtype=gs.ti_float, shape=self._solver._batch_shape(self.n_qs))
+        self.struct_rrt_node_info = ti.types.struct(
+            configuration=ti.types.vector(self.n_qs, gs.ti_float),
+            parent_idx=gs.ti_int,
+            child_idx=gs.ti_int,
+        )
+        self._rrt_node_info = self.struct_rrt_node_info.field(shape=self._solver._batch_shape(self._rrt_max_nodes))
+        self._rrt_tree_size = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        self._rrt_is_active = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        self._rrt_goal_reached_node_idx = ti.field(dtype=gs.ti_int, shape=self._solver._batch_shape())
+        
+    def _reset_rrt_connect_fields(self):
+        self._rrt_start_configuration.fill(0.0)
+        self._rrt_goal_configuration.fill(0.0)
+        self._rrt_node_info.parent_idx.fill(-1)
+        self._rrt_node_info.child_idx.fill(-1)
+        self._rrt_node_info.configuration.fill(0.0)
+        self._rrt_tree_size.fill(0)
+        self._rrt_is_active.fill(0)
+        self._rrt_goal_reached_node_idx.fill(-1)
+    
+    @ti.kernel
+    def _kernel_rrt_connect_init(
+        self,
+        qpos_start: ti.types.ndarray(),
+        qpos_goal: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray()
+    ):
+        # NOTE: run IK before this
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b_ in range(envs_idx.shape[0]):
+            i_b = envs_idx[i_b_]
+            for i_q in ti.static(range(self.n_qs)):
+                # save original qpos
+                self._rrt_start_configuration[i_q, i_b] = qpos_start[i_b_, i_q]
+                self._rrt_goal_configuration[i_q, i_b] = qpos_goal[i_b_, i_q]
+                self._rrt_node_info[0, i_b].configuration[i_q] = qpos_start[i_b_, i_q]
+                self._rrt_node_info[1, i_b].configuration[i_q] = qpos_goal[i_b_, i_q]
+            self._rrt_node_info[0, i_b].parent_idx = 0
+            self._rrt_node_info[1, i_b].child_idx = 1
+            self._rrt_tree_size[i_b] = 2
+            self._rrt_is_active[i_b] = 1
+
+    @ti.kernel
+    def _kernel_rrt_connect_cleanup(
+        self,
+        qpos_current: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray()
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b_ in range(envs_idx.shape[0]):
+            i_b = envs_idx[i_b_]
+            for i_q in ti.static(range(self.n_qs)):
+                self._solver.qpos[i_q + self._q_start, i_b] = qpos_current[i_b_, i_q]
+            self._solver._func_forward_kinematics_entity(self._entity._idx_in_solver, i_b)
+    
+    @ti.kernel
+    def _kernel_rrt_connect_step1(
+        self,
+        forward_pass: ti.i32,
+        q_limit_lower: ti.types.ndarray(),
+        q_limit_upper: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray(),
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b in envs_idx:
+            if not self._rrt_is_active[i_b]:
+                continue
+            
+            random_sample = ti.Vector([
+                q_limit_lower[i_q] + ti.random(dtype=gs.ti_float) * (q_limit_upper[i_q] - q_limit_lower[i_q])
+                for i_q in ti.static(range(self.n_qs))
+            ])
+            if ti.random() < self._rrt_goal_bias:
+                if forward_pass:
+                    random_sample = ti.Vector([
+                        self._rrt_goal_configuration[i_q, i_b]
+                        for i_q in ti.static(range(self.n_qs))
+                    ])
+                else:
+                    random_sample = ti.Vector([
+                        self._rrt_start_configuration[i_q, i_b]
+                        for i_q in ti.static(range(self.n_qs))
+                    ])
+
+            # find nearest neighbor
+            nearest_neighbor_idx = -1
+            nearest_neighbor_dist = gs.ti_float(1e30) 
+            ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+            for i_n in range(self._rrt_tree_size[i_b]):
+                if forward_pass:
+                    # NOTE: in forward pass, we only consider the previous forward pass nodes (which has parent_idx != -1)
+                    if self._rrt_node_info[i_n, i_b].parent_idx == -1:
+                        continue
+                else:
+                    # NOTE: in backward pass, we only consider the previous backward pass nodes (which has child_idx != -1)
+                    if self._rrt_node_info[i_n, i_b].child_idx == -1:
+                        continue
+                dist = (self._rrt_node_info.configuration[i_n, i_b] - random_sample).norm()
+                if dist < nearest_neighbor_dist:
+                    nearest_neighbor_dist = dist
+                    nearest_neighbor_idx = i_n
+            
+            # steer from nearest neighbor to random sample
+            nearest_config = self._rrt_node_info.configuration[nearest_neighbor_idx, i_b]
+            direction = random_sample - nearest_config
+            steer_result = ti.Vector.zero(gs.ti_float, self.n_qs)
+            for i_q in ti.static(range(self.n_qs)):
+                # If the step size exceeds max_step_size, clip it
+                if abs(direction[i_q]) > self._rrt_max_step_size:
+                    direction[i_q] = ti.math.sign(direction[i_q]) * self._rrt_max_step_size
+                steer_result[i_q] = nearest_config[i_q] + direction[i_q]
+
+            if self._rrt_tree_size[i_b] < self._rrt_max_nodes - 1:
+                # add new node
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].configuration = steer_result
+                if forward_pass:
+                    self._rrt_node_info[self._rrt_tree_size[i_b], i_b].parent_idx = nearest_neighbor_idx
+                else:
+                    self._rrt_node_info[self._rrt_tree_size[i_b], i_b].child_idx = nearest_neighbor_idx
+                self._rrt_tree_size[i_b] += 1
+
+                # set the steer result and collision check for i_b
+                for i_q in ti.static(range(self.n_qs)):
+                    self._solver.qpos[i_q + self._q_start, i_b] = steer_result[i_q]
+                self._solver._func_forward_kinematics_entity(self._entity._idx_in_solver, i_b)
+                self._solver._func_update_geoms(i_b)
+
+    @ti.kernel
+    def _kernel_rrt_connect_step2(
+        self,
+        forward_pass: ti.i32,
+        ignore_geom_pairs: ti.types.ndarray(),
+        envs_idx: ti.types.ndarray(),
+    ):
+        ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+        for i_b in envs_idx:
+            if not self._rrt_is_active[i_b]:
+                continue
+
+            collision_detected = self._func_check_collision(ignore_geom_pairs, i_b)
+            if collision_detected:
+                self._rrt_tree_size[i_b] -= 1
+                self._rrt_node_info[self._rrt_tree_size[i_b], i_b].configuration = 0.0
+                if forward_pass:
+                    self._rrt_node_info[self._rrt_tree_size[i_b], i_b].parent_idx = -1
+                else:
+                    self._rrt_node_info[self._rrt_tree_size[i_b], i_b].child_idx = -1
+            else:
+                # check the obtained steer result is within goal configuration only if no collision
+                ti.loop_config(serialize=self._solver._para_level < gs.PARA_LEVEL.ALL)
+                for i_n in range(self._rrt_tree_size[i_b]):
+                    if forward_pass:
+                        # NOTE: in forward pass, we only consider the previous backward pass nodes (which has child_idx != -1)
+                        if self._rrt_node_info[i_n, i_b].child_idx == -1:
+                            continue
+                    else:
+                        # NOTE: in backward pass, we only consider the previous forward pass nodes (which has parent_idx != -1)
+                        if self._rrt_node_info[i_n, i_b].parent_idx == -1:
+                            continue
+                    flag = True
+                    for i_q in range(self.n_qs):
+                        if abs(self._solver.qpos[i_q + self._q_start, i_b] - self._rrt_node_info.configuration[i_n, i_b][i_q]) > self._rrt_max_step_size:
+                            flag = False
+                            break
+                    if flag:
+                        self._rrt_goal_reached_node_idx[i_b] = self._rrt_tree_size[i_b] - 1
+                        if forward_pass:
+                            self._rrt_node_info[self._rrt_goal_reached_node_idx[i_b], i_b].child_idx = i_n
+                        else:
+                            self._rrt_node_info[self._rrt_goal_reached_node_idx[i_b], i_b].parent_idx = i_n
+                        self._rrt_is_active[i_b] = 0
+                        break
+        
+    def plan(
+        self,
+        qpos_goal,
+        qpos_start=None,
+        resolution=0.01,
+        timeout=5.0,
+        max_retry=1,
+        smooth_path=True,
+        num_waypoints=100,
+        ignore_collision=False,
+        envs_idx=None
+    ):
+        """
+        Plan a path from `qpos_start` to `qpos_goal`.
+        qpos_goal : array_like
+            The goal state. [B, Nq] or [1, Nq]
+        qpos_start : None | array_like, optional
+            The start state. If None, the current state of the rigid entity will be used. Defaults to None. [B, Nq] or [1, Nq]
+        """
+        import time
+        if self._solver.n_envs > 0:
+            envs_idx = self._solver._sanitize_envs_idx(envs_idx)
+        else:
+            envs_idx = torch.zeros(1, dtype=gs.tc_int, device=gs.device)
+
+        qpos_cur = self._entity.get_qpos()
+        if qpos_start is None:
+            qpos_start = qpos_cur.clone()
+        else:
+            if qpos_start.ndim == 1:
+                qpos_start = qpos_start.unsqueeze(0)
+            assert qpos_start.ndim == 2
+            if qpos_start.shape[0] == 1:
+                qpos_start = qpos_start.repeat(len(envs_idx), 1)
+
+        if qpos_goal.ndim == 1:
+            qpos_goal = qpos_goal.unsqueeze(0)
+        assert qpos_goal.ndim == 2
+        if qpos_goal.shape[0] == 1:
+            qpos_goal = qpos_goal.repeat(len(envs_idx), 1)
+        else:
+            assert qpos_goal.shape[0] == len(envs_idx), f"Batch size mismatch: {qpos_goal.shape=} {len(envs_idx)=}"
+        
+        self._init_rrt_connect_fields()
+        self._reset_rrt_connect_fields()
+        self._kernel_rrt_connect_init(qpos_start.contiguous(), qpos_goal.contiguous(), envs_idx)
+
+        self._entity.set_qpos(qpos_start)
+        self._solver._kernel_detect_collision()
+        scene_contact_info = self._solver.collider.contact_data.to_torch(gs.device)
+        n_contacts = self._solver.collider.n_contacts.to_torch(gs.device)
+
+        valid_mask = torch.logical_or(
+            torch.logical_and(
+                scene_contact_info["geom_a"] >= self.geom_start,
+                scene_contact_info["geom_a"] < self.geom_end,
+            ),
+            torch.logical_and(
+                scene_contact_info["geom_b"] >= self.geom_start,
+                scene_contact_info["geom_b"] < self.geom_end,
+            ),
+        )
+        contact_indices = torch.arange(valid_mask.shape[0], device=valid_mask.device).unsqueeze(1)
+        valid_mask = torch.logical_and(valid_mask, contact_indices < n_contacts)
+
+        max_env_collisions = int(torch.max(n_contacts).item())
+        valid_mask = valid_mask[:max_env_collisions]
+        geom_a = scene_contact_info["geom_a"][:max_env_collisions][valid_mask] # N 
+        geom_b = scene_contact_info["geom_b"][:max_env_collisions][valid_mask] # N
+        assert len(geom_a) == len(geom_b)
+
+        # NOTE: we will reduce the contacts in batch dim assuming internal geom collisions are the same for a robot
+        stacked_tensors = torch.stack((geom_a, geom_b), dim=1)
+        unique_pairs = torch.unique(stacked_tensors, dim=0) # N', 2
+
+        gs.logger.info("start rrt connect planning...")
+        start = time.time()
+        forward_pass = True
+        for i_n in range(self._rrt_max_nodes):
+            if self._rrt_is_active.to_torch().any():
+                self._kernel_rrt_connect_step1(
+                    forward_pass=forward_pass,
+                    q_limit_lower=self._entity.q_limit[0],
+                    q_limit_upper=self._entity.q_limit[1],
+                    envs_idx=envs_idx,
+                )
+                self._solver._kernel_detect_collision()
+                self._kernel_rrt_connect_step2(
+                    forward_pass=forward_pass,
+                    ignore_geom_pairs=unique_pairs,
+                    envs_idx=envs_idx,
+                )
+                forward_pass = not forward_pass
+            else:
+                break
+            
+        if self._rrt_is_active.to_torch().any():
+            gs.logger.warning(f"rrt connect planning failed in {self._rrt_is_active.to_torch().sum()} envs")
+
+        gs.logger.info(f"rrt connect planning time: {time.time() - start}")
+        start = time.time()
+
+        ts = self._rrt_tree_size.to_torch(device=gs.device)[envs_idx]
+        g_n = self._rrt_goal_reached_node_idx.to_torch(device=gs.device)[envs_idx] # B
+
+        node_info = self._rrt_node_info.to_torch(device=gs.device)
+        parents_idx = node_info["parent_idx"][:, envs_idx]
+        children_idx = node_info["child_idx"][:, envs_idx]
+        configurations = node_info["configuration"][:, envs_idx]
+        res = [g_n]
+        for _ in range(ts.max() // 2):
+            g_n = parents_idx[g_n, torch.arange(len(envs_idx))]
+            res.append(g_n)
+            if torch.all(g_n == 0):
+                break
+        res_idx = torch.stack(list(reversed(res)), dim=0)
+
+        c_n = self._rrt_goal_reached_node_idx.to_torch(device=gs.device)[envs_idx] # B
+        res = []
+        for _ in range(ts.max() // 2):
+            c_n = children_idx[c_n, torch.arange(len(envs_idx))]
+            res.append(c_n)
+            if torch.all(c_n == 1):
+                break
+        res_idx = torch.cat([res_idx, torch.stack(res, dim=0)], dim=0)
+
+        sol = configurations[res_idx, torch.arange(len(envs_idx))] # N, B, DoF
+        mask = rrt_connect_valid_mask(res_idx)
+        sol = align_weypoints_length(sol, mask, mask.sum(dim=0).max())
+
+        sol = self.shortcut_path(
+            torch.ones_like(sol[...,0]), sol, iterations=10, ignore_geom_pairs=unique_pairs, envs_idx=envs_idx
+        )
+
+        # interpolate to make num_waypoints
+        sol = align_weypoints_length(sol, torch.ones_like(sol[...,0]).bool(), num_waypoints)
+        gs.logger.info(f"path post-processing time: {time.time() - start}")
+
+        self._kernel_rrt_connect_cleanup(qpos_cur.contiguous(), envs_idx)
+        return sol
 
 
 # ------------------------------------------------------------------------------------
@@ -262,15 +871,11 @@ class OMPL(PathPlanner):
     planner : str, optional
         The name of the motion planning algorithm to use. Defaults to 'RRTConnect'.
         Supported planners: 'PRM', 'RRT', 'RRTConnect', 'RRTstar', 'EST', 'FMT', 'BITstar', 'ABITstar'.
-    resolution : float, optiona
-        Joint-space resolution in pourcentage. It corresponds to the maximum distance between states to be checked
-        for validity along a path segment. Default to 1%.
     """
     def __init__(
         self,
         entity,
         planner: Literal["RRTConnect", "RRT", "PRM", "EST", "FMT", "BITstar", "ABITstar"] = "RRTConnect",
-        resolution=0.01,
     ):
         super().__init__(entity)
         if not IS_OMPL_AVAILABLE:
@@ -278,45 +883,19 @@ class OMPL(PathPlanner):
                 gs.raise_exception_from("No pre-compiled binaries of OMPL are distributed on Windows OS.")
             else:
                 gs.raise_exception("OMPL is not installed. Please install OMPL to use this planner.")
-        
-        ou.setLogLevel(ou.LOG_ERROR)
-        self.space = ob.RealVectorStateSpace(self.n_qs)
-        bounds = ob.RealVectorBounds(self.n_qs)
-
-        for i_q in range(self.n_qs):
-            bounds.setLow(i_q, self.default_q_limit[0][i_q])
-            bounds.setHigh(i_q, self.default_q_limit[1][i_q])
-        self.space.setBounds(bounds)
-        self.ss = og.SimpleSetup(self.space)
-
-        self.si = self.ss.getSpaceInformation()
-        self.si.setStateValidityCheckingResolution(resolution)
-
-        def allocOBValidStateSampler(si):
-            vss = ob.UniformValidStateSampler(si)
-            vss.setNrAttempts(100)
-            return vss
-
-        self.si.setValidStateSamplerAllocator(ob.ValidStateSamplerAllocator(allocOBValidStateSampler))
-
-        try:
-            planner_cls = getattr(og, planner)
-            if not issubclass(planner_cls, ob.Planner):
-                raise ValueError
-            planner = planner_cls(self.si)
-        except (AttributeError, ValueError) as e:
-            gs.raise_exception_from(f"'{planner}' is not a valid planner. See OMPL documentation for details.", e)
-        self.ss.setPlanner(planner)
+        self.planner = planner
     
     def plan(
         self,
         qpos_goal,
         qpos_start=None,
         *,
+        resolution=0.01,
         timeout=5.0,
         max_retry=1,
         smooth_path=True,
         num_waypoints=100,
+        ignore_joint_limit:bool=False,
         ignore_collision:bool=False
     ) -> torch.Tensor:
         """
@@ -328,6 +907,9 @@ class OMPL(PathPlanner):
             The goal state.
         qpos_start : None | array_like, optional
             The start state. If None, the current state of the rigid entity will be used. Defaults to None.
+        resolution : float, optiona
+            Joint-space resolution in pourcentage. It corresponds to the maximum distance between states to be checked
+            for validity along a path segment. Default to 1%.
         timeout : float, optional
             The maximum time (in seconds) allowed for the motion planning algorithm to find a solution. Defaults to 5.0.
         max_retry : float, optional
@@ -342,12 +924,6 @@ class OMPL(PathPlanner):
         waypoints : torch.Tensor
             A list of waypoints representing the planned path. Each waypoint is an array storing the entity's qpos of a single time step.
         """
-        if not IS_OMPL_AVAILABLE:
-            if gs.platform == "Windows":
-                gs.raise_exception_from("No pre-compiled binaries of OMPL are not distributed on Windows OS.")
-            else:
-                gs.raise_exception_from("OMPL not found.")
-
         assert timeout > 0.0 and math.isfinite(timeout)
         assert max_retry > 0
 
@@ -356,6 +932,44 @@ class OMPL(PathPlanner):
 
         if self.n_qs != self.n_dofs:
             gs.raise_exception("Motion planning is not yet supported for rigid entities with free joints.")
+
+        if qpos_start is None:
+            qpos_start = self._entity.get_qpos()
+        qpos_start = tensor_to_array(qpos_start)
+        qpos_goal = tensor_to_array(qpos_goal)
+
+        if qpos_start.shape != (self.n_qs,) or qpos_goal.shape != (self.n_qs,):
+            gs.raise_exception("Invalid shape for `qpos_start` or `qpos_goal`.")
+
+        ######### process joint limit ##########
+        if ignore_joint_limit:
+            gs.logger.warning("This option is deprecated and is no longer doing anything.")
+        q_limit_lower, q_limit_upper = self._entity.q_limit[0], self._entity.q_limit[1]
+
+        if (qpos_start < q_limit_lower).any() or (qpos_start > q_limit_upper).any():
+            gs.logger.warning(
+                "`qpos_start` exceeds joint limit. Relaxing joint limit to contain `qpos_start` for planning."
+            )
+            q_limit_lower = np.minimum(q_limit_lower, qpos_start)
+            q_limit_upper = np.maximum(q_limit_upper, qpos_start)
+
+        if (qpos_goal < q_limit_lower).any() or (qpos_goal > q_limit_upper).any():
+            gs.logger.warning(
+                "`qpos_goal` exceeds joint limit. Relaxing joint limit to contain `qpos_goal` for planning."
+            )
+            q_limit_lower = np.minimum(q_limit_lower, qpos_goal)
+            q_limit_upper = np.maximum(q_limit_upper, qpos_goal)
+
+        ######### setup OMPL ##########
+        ou.setLogLevel(ou.LOG_ERROR)
+        space = ob.RealVectorStateSpace(self.n_qs)
+        bounds = ob.RealVectorBounds(self.n_qs)
+
+        for i_q in range(self.n_qs):
+            bounds.setLow(i_q, q_limit_lower[i_q])
+            bounds.setHigh(i_q, q_limit_upper[i_q])
+        space.setBounds(bounds)
+        ss = og.SimpleSetup(space)
 
         geoms_idx = tuple(range(self.geom_start, self.geom_start + self.n_geoms))
         mask_collision_pairs = set(
@@ -372,25 +986,41 @@ class OMPL(PathPlanner):
             collision_pairs = set(map(tuple, self._entity.detect_collision()))
             return not (collision_pairs - mask_collision_pairs)
 
-        self.ss.setStateValidityChecker(ob.StateValidityCheckerFn(is_ompl_state_valid))
+        ss.setStateValidityChecker(ob.StateValidityCheckerFn(is_ompl_state_valid))
 
-        qpos_goal, qpos_start = self.validate_input_qpos(qpos_goal, qpos_start)
+        si = ss.getSpaceInformation()
+        si.setStateValidityCheckingResolution(resolution)
 
-        # setup OMPL
-        state_start = ob.State(self.space)
-        state_goal = ob.State(self.space)
+        def allocOBValidStateSampler(si):
+            vss = ob.UniformValidStateSampler(si)
+            vss.setNrAttempts(100)
+            return vss
+
+        si.setValidStateSamplerAllocator(ob.ValidStateSamplerAllocator(allocOBValidStateSampler))
+
+        try:
+            planner_cls = getattr(og, self.planner)
+            if not issubclass(planner_cls, ob.Planner):
+                raise ValueError
+            planner = planner_cls(si)
+        except (AttributeError, ValueError) as e:
+            gs.raise_exception_from(f"'{planner}' is not a valid planner. See OMPL documentation for details.", e)
+        ss.setPlanner(planner)
+
+        state_start = ob.State(space)
+        state_goal = ob.State(space)
         for i_q in range(self.n_qs):
             state_start[i_q] = float(qpos_start[i_q])
             state_goal[i_q] = float(qpos_goal[i_q])
-        self.ss.setStartAndGoalStates(state_start, state_goal)
+        ss.setStartAndGoalStates(state_start, state_goal)
 
-        # solve
+        ######### solve ##########
         waypoints = []
         for i in range(max_retry):
             # Try solve the motion planning problem
-            if self.ss.getPlanner():
-                self.ss.getPlanner().clear()
-            status = self.ss.solve(timeout)
+            if ss.getPlanner():
+                ss.getPlanner().clear()
+            status = ss.solve(timeout)
             status_type = status.getStatus()
 
             # Check if there was some unrecoverable failure
@@ -411,14 +1041,12 @@ class OMPL(PathPlanner):
 
             # Extract solution if any
             if status:
-                path = self.ss.getSolutionPath()
+                # ss.simplifySolution()
+                path = ss.getSolutionPath()
 
-                import time
-
-                start = time.time()
                 # Simplify path
                 if smooth_path:
-                    ps = og.PathSimplifier(self.si)
+                    ps = og.PathSimplifier(si)
                     try:
                         # ps.simplifyMax(path)
                         ps.partialShortcutPath(path)
@@ -430,8 +1058,6 @@ class OMPL(PathPlanner):
                 # Interpolate path
                 if num_waypoints is not None:
                     path.interpolate(num_waypoints)
-
-                print("smooting:", time.time() - start)
 
                 # Extract waypoints
                 waypoints = [
@@ -457,48 +1083,49 @@ class OMPL(PathPlanner):
         self._entity.set_qpos(qpos_start, zero_velocity=False)
 
         return waypoints
+
     
 
 class RRTConnect_OMPL(OMPL):
     """OMPL RRT-Connect planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="RRTConnect", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="RRTConnect", **kwargs)
 
 
 class RRT_OMPL(OMPL):
     """OMPL RRT planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="RRT", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="RRT", **kwargs)
 
 
 class PRM_OMPL(OMPL):
     """OMPL PRM (Probabilistic Roadmap) planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="PRM", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="PRM", **kwargs)
 
 
 class EST_OMPL(OMPL):
     """OMPL EST (Expansive-Space Tree) planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="EST", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="EST", **kwargs)
 
 
 class FMT_OMPL(OMPL):
     """OMPL FMT* (Fast Marching Tree) planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="FMT", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="FMT", **kwargs)
 
 
 class BITstar_OMPL(OMPL):
     """OMPL BIT* (Batch Informed Trees) planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="BITstar", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="BITstar", **kwargs)
 
 
 class ABITstar_OMPL(OMPL):
     """OMPL ABIT* (Asymptotically-Optimal Batch Informed Trees) planner."""
-    def __init__(self, entity, resolution: float = 0.01, **kwargs):
-        super().__init__(entity, planner="ABITstar", resolution=resolution, **kwargs)
+    def __init__(self, entity, **kwargs):
+        super().__init__(entity, planner="ABITstar", **kwargs)
 
 
 # ------------------------------------------------------------------------------------
